+++ src/Abstractions/DurableTaskRegistry.Activities.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Reflection;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.DurableTask;
@@ -246,7 +245,7 @@ public sealed partial class DurableTaskRegistry
         Func<TaskActivityContext, TInput, Task<TOutput>> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 
     /// <summary>
@@ -264,7 +263,7 @@ public sealed partial class DurableTaskRegistry
         Func<TaskActivityContext, TInput, TOutput> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 
     /// <summary>
@@ -280,7 +279,7 @@ public sealed partial class DurableTaskRegistry
     public DurableTaskRegistry AddActivityFunc<TInput>(Func<TaskActivityContext, TInput, Task> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 
     /// <summary>
@@ -296,7 +295,7 @@ public sealed partial class DurableTaskRegistry
     public DurableTaskRegistry AddActivityFunc<TOutput>(Func<TaskActivityContext, Task<TOutput>> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 
     /// <summary>
@@ -311,7 +310,7 @@ public sealed partial class DurableTaskRegistry
     public DurableTaskRegistry AddActivityFunc(Func<TaskActivityContext, Task> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 
     /// <summary>
@@ -327,7 +326,7 @@ public sealed partial class DurableTaskRegistry
     public DurableTaskRegistry AddActivityFunc<TOutput>(Func<TaskActivityContext, TOutput> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 
     /// <summary>
@@ -343,7 +342,7 @@ public sealed partial class DurableTaskRegistry
     public DurableTaskRegistry AddActivityFunc<TInput>(Action<TaskActivityContext, TInput> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 
     /// <summary>
@@ -358,39 +357,6 @@ public sealed partial class DurableTaskRegistry
     public DurableTaskRegistry AddActivityFunc(Action<TaskActivityContext> activity)
     {
         Check.NotNull(activity);
-        return this.AddActivityFunc(GetActivityNameFromDelegate(activity), activity);
-    }
-
-    /// <summary>
-    /// Gets the task name from a delegate by checking for a <see cref="DurableTaskAttribute"/>
-    /// or falling back to the method name.
-    /// </summary>
-    /// <param name="delegate">The delegate to extract the name from.</param>
-    /// <returns>The task name.</returns>
-    /// <exception cref="ArgumentException">
-    /// Thrown if the name cannot be inferred from the delegate.
-    /// </exception>
-    static TaskName GetActivityNameFromDelegate(Delegate @delegate)
-    {
-        MethodInfo method = @delegate.Method;
-
-        // Check for DurableTaskAttribute on the method
-        DurableTaskAttribute? attribute = method.GetCustomAttribute<DurableTaskAttribute>();
-        if (attribute?.Name.Name is not null and not "")
-        {
-            return attribute.Name;
-        }
-
-        // Fall back to method name
-        string? methodName = method.Name;
-        if (string.IsNullOrEmpty(methodName) || methodName.StartsWith("<", StringComparison.Ordinal))
-        {
-            throw new ArgumentException(
-                "Cannot infer activity name from the delegate. The delegate must either have a " +
-                "[DurableTask] attribute with a name, or be a named method (not a lambda or anonymous delegate).",
-                nameof(@delegate));
-        }
-
-        return new TaskName(methodName);
+        return this.AddActivityFunc(GetTaskNameFromDelegate(activity), activity);
     }
 }
