+++ src/Abstractions/DurableTaskRegistry.Orchestrators.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Reflection;
-
 namespace Microsoft.DurableTask;
 
 /// <summary>
@@ -365,38 +363,5 @@ public partial class DurableTaskRegistry
     {
         Check.NotNull(orchestrator);
         return this.AddOrchestratorFunc(GetTaskNameFromDelegate(orchestrator), orchestrator);
-    }
-
-    /// <summary>
-    /// Gets the task name from a delegate by checking for a <see cref="DurableTaskAttribute"/>
-    /// or falling back to the method name.
-    /// </summary>
-    /// <param name="delegate">The delegate to extract the name from.</param>
-    /// <returns>The task name.</returns>
-    /// <exception cref="ArgumentException">
-    /// Thrown if the name cannot be inferred from the delegate.
-    /// </exception>
-    static TaskName GetTaskNameFromDelegate(Delegate @delegate)
-    {
-        MethodInfo method = @delegate.Method;
-
-        // Check for DurableTaskAttribute on the method
-        DurableTaskAttribute? attribute = method.GetCustomAttribute<DurableTaskAttribute>();
-        if (attribute?.Name.Name is not null and not "")
-        {
-            return attribute.Name;
-        }
-
-        // Fall back to method name
-        string? methodName = method.Name;
-        if (string.IsNullOrEmpty(methodName) || methodName.StartsWith("<", StringComparison.Ordinal))
-        {
-            throw new ArgumentException(
-                "Cannot infer orchestrator name from the delegate. The delegate must either have a " +
-                "[DurableTask] attribute with a name, or be a named method (not a lambda or anonymous delegate).",
-                nameof(@delegate));
-        }
-
-        return new TaskName(methodName);
     }
 }
